feat(sn_url): moving codebase from sn_api::safe_url module as a standalone sn_url crate
This is the SafeUrl implementation which was part of https://github.com/maidsafe/sn_api.git repository
and it's now being moved out to this new repository to become a standalone crate. This first commit
ontains all the contributions made to the SafeUrl module by several developers/authors as listed below.

Co-authored-by: Josh Wilson <joshuef@gmail.com>
Co-authored-by: Ravinder Jangra <ravinderjangra@live.com>
Co-authored-by: Ian Coleman <ian@iancoleman.io>
Co-authored-by: danda <danda@osc.co.cr>
Co-authored-by: Stephen Coyle <stephen.coyle@maidsafe.net>